+++ src/components/Home.js
@@ -3,7 +3,9 @@ import { useQuery } from '@apollo/react-hooks';
 import gql from 'graphql-tag';
 
 import Loading from './Loading';
+import GraphQLError from './GraphQLError';
 import SetKey from './SetKey';
+import GuildSummary from './Guild/Summary';
 
 const CHECK_TOKEN = gql`query {
 	me {
@@ -11,20 +13,30 @@ const CHECK_TOKEN = gql`query {
 		token { id }
 		gw2 { id name }
 	}
+	gw2 {
+		guild {
+			id
+		}
+	}
 }`;
 
 function Home() {
 	const [key_key, setKeyKey] = useState(new Date().getTime());
-	const { data, loading, refetch } = useQuery(CHECK_TOKEN);
+	const { data, loading, error, refetch } = useQuery(CHECK_TOKEN);
 
 	const onExpires = useCallback(() => {
 		setKeyKey(new Date().getTime());
 	}, []);
 
 	if (loading) return <Loading />;
-	if (data && ! data.me.token) return <SetKey key={ key_key } onExpires={ onExpires } onSet={ refetch } />;
+	if (error) return <GraphQLError error={ error } />;
 
-	return <div>Welcome { data && data.me.gw2.name }!</div>;
+	return <div>
+		{ data.gw2.guild && <GuildSummary guild={ data.gw2.guild.id } /> }
+		{ data.me.token ? <div>
+			Welcome { data && data.me.gw2.name }!
+		</div> : <SetKey key={ key_key } onExpires={ onExpires } onSet={ refetch } /> }
+	</div>;
 }
 
 export default Home;