+++ CHANGELOG.md
@@ -5,6 +5,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased][]
 
+### Added
+- Guild summary on home page
+
 ## [1.0.0][] - 2019-10-05
 
 - Initial release